Code cleanup; A few new options
Cleaned up (deleted) some of the extraneous classes that had only 1 or 2
class methods.  Replaced these with module methods in the DNSOMatic namespace.

Added a -x/--debug option which currently outputs a backtrace when exceptions
are handled.

Added a -a/--alert option which outputs a message whenever we attempt to
update our IP (either through a normal change or forced).  This is a verbose
option that is specific to this one output.